AppleClang fixes.

+++ src/execution/sql/functions/string_functions.cpp
@@ -14,7 +14,7 @@ void StringFunctions::Substring(UNUSED exec::ExecutionContext *ctx, StringVal *r
     return;
   }
 
-  const auto start = std::max(pos.val, 1l);
+  const auto start = std::max(pos.val, static_cast<i64>(1));
   const auto end = pos.val + std::min(static_cast<i64>(str.len), len.val);
 
   // The end can be before the start only if the length was negative. This is an
@@ -343,7 +343,8 @@ void StringFunctions::Left(UNUSED exec::ExecutionContext *ctx, StringVal *result
     return;
   }
 
-  const auto len = n.val < 0 ? std::max(0l, str.len + n.val) : std::min(str.len, static_cast<u32>(n.val));
+  const auto len = n.val < 0 ? std::max(static_cast<u32>(0), static_cast<u32>(str.len + n.val))
+                             : std::min(str.len, static_cast<u32>(n.val));
   *result = StringVal(str.Content(), u32(len));
 }
